Revisi Product dan Pricing

+++ database/migrations/2026_06_23_000006_create_product_images_table.php
@@ -1,4 +1,5 @@
 <?php
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -13,7 +14,7 @@ return new class extends Migration
             $table->unsignedBigInteger('product_variant_id')->nullable();
             $table->string('image_path', 255);
             $table->string('alt_text', 150)->nullable();
-            $table->integer('sort_order')->default(0);
+            $table->unsignedSmallInteger('sort_order')->default(0);
             $table->boolean('is_primary')->default(false);
             $table->unsignedBigInteger('created_by')->nullable();
             $table->unsignedBigInteger('updated_by')->nullable();
@@ -21,7 +22,7 @@ return new class extends Migration
 
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->foreign('product_variant_id')->references('id')->on('product_variants')->onDelete('set null');
-            
+
             $table->index('sort_order');
             $table->index('is_primary');
         });
