+++ database/migrations/2026_06_23_000008_create_product_attribute_values_table.php
@@ -1,4 +1,5 @@
 <?php
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -11,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('attribute_id');
             $table->string('value', 100);
-            $table->integer('sort_order')->default(0);
+            $table->unsignedSmallInteger('sort_order')->default(0);
             $table->timestamps();
 
             $table->foreign('attribute_id')->references('id')->on('product_attributes')->onDelete('cascade');
